+++ CHANGELOG.md
@@ -33,7 +33,7 @@
 
 ### Deprecation
 - [PR #1471](https://github.com/stympy/faker/pull/1471) Add music and movies namespace [@vbrazo](https://github.com/vbrazo)
-  - Deprecates `::BackToTheFuture`, `::LordOfTheRings`, `::VForVendetta`, `::StarWars`, `::PrincessBride`, `::UmphreysMcgee`
+  - Deprecates `::BackToTheFuture`, `::Lebowski`, `::LordOfTheRings`, `::PrincessBride`, `::StarWars`, `::UmphreysMcgee`, `::VForVendetta`
 - [PR #1469](https://github.com/stympy/faker/pull/1469) Deprecate Faker::Hobbit and reorganize unreleased docs and tests [@vbrazo](https://github.com/vbrazo)
   - Deprecates `::Hobbit`
 - [PR #1431](https://github.com/stympy/faker/pull/1431) Add Faker::TvShows namespace [@SpyMaster356](https://github.com/SpyMaster356)
