--- doc/unreleased/default/lebowski.md
@@ -1,11 +1,0 @@
-# Faker::Lebowski
-
-Available since version 1.8.8.
-
-```ruby
-Faker::Lebowski.actor #=> "John Goodman"
-
-Faker::Lebowski.character #=> "Jackie Treehorn"
-
-Faker::Lebowski.quote #=> "Forget it, Donny, you're out of your element!"
-```
