+++ unreleased_README.md
@@ -171,7 +171,6 @@ gem 'faker', :git => 'https://github.com/stympy/faker.git', :branch => 'master'
   - [Faker::Job](doc/unreleased/default/job.md)
   - [Faker::Json](doc/unreleased/default/json.md)
   - [Faker::Kpop](doc/unreleased/default/kpop.md)
-  - [Faker::Lebowski](doc/unreleased/default/lebowski.md)
   - [Faker::Lorem](doc/unreleased/default/lorem.md)
   - [Faker::LoremFlickr](doc/unreleased/default/lorem_flickr.md)
   - [Faker::LoremPixel](doc/unreleased/default/lorem_pixel.md)
@@ -249,6 +248,7 @@ gem 'faker', :git => 'https://github.com/stympy/faker.git', :branch => 'master'
   - [Faker::Movies::BackToTheFuture](doc/unreleased/movies/back_to_the_future.md)
   - [Faker::Movies::Ghostbusters](doc/unreleased/movies/ghostbusters.md)
   - [Faker::Movies::GratefulDead](doc/unreleased/movies/grateful_dead.md)
+  - [Faker::Movies::Lebowski](doc/unreleased/movies/lebowski.md)
   - [Faker::Movies::LordOfTheRings](doc/unreleased/movies/lord_of_the_rings.md)
   - [Faker::Movies::PrincessBride](doc/unreleased/movies/princess_bride.md)
   - [Faker::Movies::StarWars](doc/unreleased/movies/star_wars.md)
